+++ src/main/java/com/klikli_dev/theurgy/datagen/book/SpagyricsCategoryProvider.java
@@ -16,8 +16,7 @@ import com.klikli_dev.modonomicon.api.datagen.book.page.BookMultiblockPageModel;
 import com.klikli_dev.modonomicon.api.datagen.book.page.BookSpotlightPageModel;
 import com.klikli_dev.modonomicon.api.datagen.book.page.BookTextPageModel;
 import com.klikli_dev.theurgy.Theurgy;
-import com.klikli_dev.theurgy.integration.modonomicon.page.BookAccumulationRecipePageModel;
-import com.klikli_dev.theurgy.integration.modonomicon.page.BookCalcinationRecipePageModel;
+import com.klikli_dev.theurgy.integration.modonomicon.page.accumulation.BookAccumulationRecipePageModel;
 import com.klikli_dev.theurgy.registry.ItemRegistry;
 import net.minecraft.world.item.crafting.Ingredient;
 import net.minecraftforge.common.data.LanguageProvider;
