+++ src/main/java/com/klikli_dev/theurgy/integration/modonomicon/page/accumulation/BookAccumulationRecipePageModel.java
@@ -1,18 +1,18 @@
 /*
- * SPDX-FileCopyrightText: 2022 klikli-dev
+ * SPDX-FileCopyrightText: 2023 klikli-dev
  *
  * SPDX-License-Identifier: MIT
  */
 
-package com.klikli_dev.theurgy.integration.modonomicon.page;
+package com.klikli_dev.theurgy.integration.modonomicon.page.accumulation;
 
 import com.klikli_dev.modonomicon.api.datagen.book.page.BookRecipePageModel;
 import com.klikli_dev.theurgy.integration.modonomicon.TheurgyModonomiconConstants;
 import org.jetbrains.annotations.NotNull;
 
-public class BookCalcinationRecipePageModel extends BookRecipePageModel {
-    protected BookCalcinationRecipePageModel(@NotNull String anchor) {
-        super(TheurgyModonomiconConstants.Page.CALCINATION_RECIPE, anchor);
+public class BookAccumulationRecipePageModel extends BookRecipePageModel {
+    protected BookAccumulationRecipePageModel(@NotNull String anchor) {
+        super(TheurgyModonomiconConstants.Page.ACCUMULATION_RECIPE, anchor);
     }
 
     public static Builder builder() {
@@ -24,8 +24,8 @@ public class BookCalcinationRecipePageModel extends BookRecipePageModel {
             super();
         }
 
-        public BookCalcinationRecipePageModel build() {
-            var model = new BookCalcinationRecipePageModel(this.anchor);
+        public BookAccumulationRecipePageModel build() {
+            var model = new BookAccumulationRecipePageModel(this.anchor);
             model.title1 = this.title1;
             model.recipeId1 = this.recipeId1;
             model.title2 = this.title2;
